# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sdr-- API",
    "description": "",
    "contact": {
      "name": "sdr-- contributors"
    },
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/devicesets": {
      "post": {
        "operationId": "create_device_set",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeviceSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Device set created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedId"
                }
              }
            }
          },
          "400": {
            "description": "Unusable device, or one a device set already holds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Device not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Malformed request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/devicesets/{ds}/channels/{ch}/record": {
      "post": {
        "operationId": "record_channel_audio",
        "parameters": [
          {
            "name": "ds",
            "in": "path",
            "description": "Device set id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "ch",
            "in": "path",
            "description": "Channel id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelRecordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Recording status: live after `start`; final counts after `stop`, where `error` reports a recording that was cut short and the finished file appears in `GET /api/audiorecordings`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioRecordingStatus"
                }
              }
            }
          },
          "400": {
            "description": "Cannot record: no recordings directory, set not running, channel already recording, not recording, or a channel with no audio",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Device set or channel not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Malformed request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/devicesets/{ds}/device": {
      "patch": {
        "operationId": "patch_device",
        "parameters": [
          {
            "name": "ds",
            "in": "path",
            "description": "Device set id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Settings applied"
          },
          "400": {
            "description": "Unsupported setting",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Device set not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Malformed request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "AudioRecordingStatus": {
        "type": "object",
        "required": [
          "file",
          "started_at",
          "channels",
          "frames",
          "bytes"
        ],
        "properties": {
          "bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "channels": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "file": {
            "type": "string"
          },
          "frames": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "started_at": {
            "type": "string"
          }
        }
      },
...
      "ChannelRecordRequest": {
        "type": "object",
        "required": [
          "action"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/RecordAction"
          }
        }
      },
...
      "CreateDeviceSetRequest": {
        "type": "object",
        "required": [
          "device_id"
        ],
        "properties": {
          "device_id": {
            "type": "string"
          }
        }
      },
...
      "CreatedId": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "DeviceSettings": {
        "type": "object",
        "properties": {
          "antenna": {
            "type": [
              "string",
              "null"
            ]
          },
          "bandwidth": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "center_hz": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "extra": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExtraValue"
            }
          },
          "gains": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GainValue"
            }
          },
          "ppm": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "sample_rate": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "streams": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StreamSettings"
            }
          }
        }
      },
...
      "ExtraValue": {
        "type": "object",
        "required": [
          "name",
          "value"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {}
        }
      },
...
      "GainValue": {
        "type": "object",
        "required": [
          "stage",
          "value_db"
        ],
        "properties": {
          "stage": {
            "type": "string"
          },
          "value_db": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "RecordAction": {
        "type": "string",
        "enum": [
          "start",
          "stop"
        ]
      },
...
      "StreamSettings": {
        "type": "object",
        "required": [
          "stream"
        ],
        "properties": {
          "antenna": {
            "type": [
              "string",
              "null"
            ]
          },
          "center_hz": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "gains": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GainValue"
            }
          },
          "stream": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
    }
  }
}